requirements.txt
fastapi
uvicorn
tensorflow
pillow
transformers
torch
opencv-python
tensorflow-hub